11th Place/requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
numpy
lightgbm
pandas
pip
matplotlib
scikit-learn
seaborn

# backwards compatibility
pathlib2
